Reuse custom importer for both test and build styles scripts
